feat: add UserVerification + move UserCredential to authentication context
