+++ src/modules/identity/domain/interfaces/repositories/user-credentials-repository.ts
@@ -1,4 +1,4 @@
-import { UserCredential } from "../../entities/UserCredential";
+import { UserCredential } from "../../../authentication/domain/UserCredential";
 
 export abstract class UserCredentialsRepository {
   abstract add(user: UserCredential, tx?: unknown): Promise<void>;
